+++ src/eventhandlers/uinputeventhandler.cpp
@@ -295,8 +295,8 @@ void UInputEventHandler::sendMouseSpringEvent(int xDis, int yDis,
 {
     if ((width > 0) && (height > 0))
     {
-        double midwidth = static_cast<double>(width) / 2.0;
-        double midheight = static_cast<double>(height) / 2.0;
+        double midwidth = width / 2.0;
+        double midheight = height / 2.0;
 
         int fx = static_cast<int>(ceil(32767 * ((xDis - midwidth) / midwidth)));
         int fy = static_cast<int>(ceil(32767 * ((yDis - midheight) / midheight)));
